+++ app-frontend/src/routes/FlowChart/Legend.css
@@ -1,4 +1,4 @@
-@import "./FlowNode.css";
+@import "FlowNode.css";
 
 .legend {
   width: fit-content;
